pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Sebastian"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tensorflow = "^2.16.1"
keras = "^3.3.3"
pandas = "^2.2.2"
numpy = "^1.26.4"
jupyter = "^1.0.0"
notebook = "^7.1.3"
tqdm = "^4.66.4"
opencv-contrib-python = "^4.9.0.80"
scikit-learn = "^1.4.2"
pillow = "^10.3.0"
flask = "^3.0.3"
opencv-python = "^4.9.0.80"
matplotlib = "^3.9.0"
pydantic = "^2.7.2"
firebase-admin = "^6.5.0"
firebase-functions = "^0.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
